tracee-ebpf: single vmlinux header file for CO-RE
Split vmlinux header file into 2 files:

1. vmlinux/vmlinux.h: containing CO-RE capable type definitions

2. vmlinux/vmlinux-flavors.h: included at the end of vmlinux.h file and
   supporting different types flavors for CO-RE run-time relocations to
   work.
