+++ tracee-ebpf/tracee/tracee.bpf.c
@@ -64,7 +64,7 @@ Copyright (C) Aqua Security inc.
 
 #else
 //CO:RE is enabled
-#include <vmlinux-arch.h>
+#include <vmlinux.h>
 #include "co_re_missing_definitions.h"
 
 #endif
